backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
pydantic-settings
python-dotenv
python-multipart
